Deleted most of what I had done. dv no longer takes src
Had to also fix my nodesNeighbors list maker because for node w, both u and v had the same value, so I had to not use the index() function and instead made an outside variable that would keep track of val's index, and this worked perfectly.
So I had a list of lists of all the neighbors of all the nodes, now just have to figure out how to get the distance vector for each?
Figured out how to calculate bellman-ford. Maybe use a function for recursion? Need help xD