Changed a thing or two and added comments
Main change was that now I have made allValues an outside variable to keep track of all the values of all the nodes, but without the node's letter. I guess this could also have been done with topology[i][1:] but I did it my own way since I was already going through all the values of all the nodes to find who is a neighbor and who is not anyway.

So now I have a list of all the nodes without an empty string before it, a list of lists of all their values corresponding to their indexes, and a list of lists of all their neighbors. I think with this, we can create a function to calculate Bellman-Ford, because I cannot wrap my head around a way to do it without recursion.

For instance, most of the code is inside a for loop, the first iteration is the first node, let's say node U. Now all I have all the values and neighbors for U, but I cannot calculate any distance vectors to a node that is not a neighbor of U because I only know the neighbors of U at the end of this for loop.
Hence why I think we might need a function after getting everything so that I know everything already and I just need to grab it. So I would start u to u = 0, u to v = 7, u to w = 3, u to x = 5, then u to y = ??? well, I'd have to call the function again for v, w, and x and see which one of them is neighbors with y or keep calling the function until a neighbor of Y is found, which at the end would return the cost found.
Still a lot to wrap my head around, and how would I even begin to keep track of the cost if we get a 30x30 topology?